Rename client to client_jobs
